requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
requests==2.23.0
six==1.15.0
urllib3==1.25.9
python-status==1.0.1
